+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 
 - Verbosity argument accepts level (DEBUG, INFO, WARNING, ERROR, CRITICAL)
 - log_level setting in site section of site.conf
+- Plugin site_search provides json index and javascript to implement a basic site search